backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
openai==1.6.1
apscheduler==3.10.4
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
PyYAML==6.0.1
python-slugify==8.0.1
markdown==3.5.1
gunicorn==21.2.0
